backend/requirements.txt
Flask
flask-cors
requests
python-dotenv
Flask-SocketIO
